// src/app/Modules/core/models/Oficina.ts
// Generated by https://quicktype.io

export interface Oficina {
  office_id:    number;
  office_level: number;
  office_name:  string;
  address:      string;
  country:      string;
  city:         string;
  phone:        string;
  email:        string;
  office_code:  string;
  office_dep:   string;
  status:       number;
}

export interface OficinaUi extends Oficina {
  hijos : OficinaUi[];
}
// Generated by https://quicktype.io

export interface OficinaPost extends Omit<Oficina, 'office_id' |'office_code' > {

}