// src/app/Modules/core/models/PolizaExtra.model.ts
// Generated by https://quicktype.io

export interface PolizaExtra {
  poliza_extra_id: number;
  venta_id:        number;
  beneficio_id:    number;
  monto_adicional: number;
}


// Generated by https://quicktype.io

export interface PolizaExtraToPost {
  venta_id:        number;
  beneficio_id:    string;
  monto_adicional: string;
}